Allow easier application debugging (#3)
* Add .gitignore

Exclude generated binaries from git status with a basic .gitignore

Tests:
 - Build boot loader and updater and verify no additional files
   are marked as untracked in "git status"

* Enable debug symbols

Allow debugging in gdb and assembler analysis (if
required) by turning on the generation of debug
symbols during the build. Use the -ggdb3 option in
preference to -g as this generates the most
expressive debugging information for GNU tools
which are the only likely users of the symbols.